+++ types/dec_coin.go
@@ -575,7 +575,6 @@ func removeZeroDecCoins(coins DecCoins) DecCoins {
 
 var _ sort.Interface = Coins{}
 
-//nolint
 func (coins DecCoins) Len() int           { return len(coins) }
 func (coins DecCoins) Less(i, j int) bool { return coins[i].Denom < coins[j].Denom }
 func (coins DecCoins) Swap(i, j int)      { coins[i], coins[j] = coins[j], coins[i] }