+++ x/ibc/03-connection/keeper/keeper.go
@@ -20,14 +20,16 @@ import (
 // Keeper defines the IBC connection keeper
 type Keeper struct {
 	storeKey     sdk.StoreKey
-	cdc          *codec.Codec
+	aminoCdc     *codec.Codec    // amino codec. TODO: remove after clients have been migrated to proto
+	cdc          codec.Marshaler // hybrid codec
 	clientKeeper types.ClientKeeper
 }
 
 // NewKeeper creates a new IBC connection Keeper instance
-func NewKeeper(cdc *codec.Codec, key sdk.StoreKey, ck types.ClientKeeper) Keeper {
+func NewKeeper(aminoCdc *codec.Codec, cdc codec.Marshaler, key sdk.StoreKey, ck types.ClientKeeper) Keeper {
 	return Keeper{
 		storeKey:     key,
+		aminoCdc:     aminoCdc,
 		cdc:          cdc,
 		clientKeeper: ck,
 	}
@@ -54,13 +56,14 @@ func (k Keeper) GetConnection(ctx sdk.Context, connectionID string) (types.Conne
 
 	var connection types.ConnectionEnd
 	k.cdc.MustUnmarshalBinaryBare(bz, &connection)
+
 	return connection, true
 }
 
 // SetConnection sets a connection to the store
 func (k Keeper) SetConnection(ctx sdk.Context, connectionID string, connection types.ConnectionEnd) {
 	store := ctx.KVStore(k.storeKey)
-	bz := k.cdc.MustMarshalBinaryBare(connection)
+	bz := k.cdc.MustMarshalBinaryBare(&connection)
 	store.Set(ibctypes.KeyConnection(connectionID), bz)
 }
 
@@ -90,15 +93,16 @@ func (k Keeper) GetClientConnectionPaths(ctx sdk.Context, clientID string) ([]st
 		return nil, false
 	}
 
-	var paths []string
-	k.cdc.MustUnmarshalBinaryBare(bz, &paths)
-	return paths, true
+	var clientPaths types.ClientPaths
+	k.cdc.MustUnmarshalBinaryBare(bz, &clientPaths)
+	return clientPaths.Paths, true
 }
 
 // SetClientConnectionPaths sets the connections paths for client
 func (k Keeper) SetClientConnectionPaths(ctx sdk.Context, clientID string, paths []string) {
 	store := ctx.KVStore(k.storeKey)
-	bz := k.cdc.MustMarshalBinaryBare(paths)
+	clientPaths := types.ClientPaths{Paths: paths}
+	bz := k.cdc.MustMarshalBinaryBare(&clientPaths)
 	store.Set(ibctypes.KeyClientConnections(clientID), bz)
 }