+++ x/ibc/07-tendermint/types/tendermint_test.go
@@ -6,11 +6,13 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
+	abci "github.com/tendermint/tendermint/abci/types"
 	tmtypes "github.com/tendermint/tendermint/types"
 
 	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/simapp"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	ibctmtypes "github.com/cosmos/cosmos-sdk/x/ibc/07-tendermint/types"
-	commitmenttypes "github.com/cosmos/cosmos-sdk/x/ibc/23-commitment/types"
 )
 
 const (
@@ -24,18 +26,21 @@ const (
 type TendermintTestSuite struct {
 	suite.Suite
 
-	cdc     *codec.Codec
-	privVal tmtypes.PrivValidator
-	valSet  *tmtypes.ValidatorSet
-	header  ibctmtypes.Header
-	now     time.Time
+	ctx      sdk.Context
+	aminoCdc *codec.Codec
+	cdc      codec.Marshaler
+	privVal  tmtypes.PrivValidator
+	valSet   *tmtypes.ValidatorSet
+	header   ibctmtypes.Header
+	now      time.Time
 }
 
 func (suite *TendermintTestSuite) SetupTest() {
-	suite.cdc = codec.New()
-	codec.RegisterCrypto(suite.cdc)
-	ibctmtypes.RegisterCodec(suite.cdc)
-	commitmenttypes.RegisterCodec(suite.cdc)
+	checkTx := false
+	app := simapp.Setup(checkTx)
+
+	suite.aminoCdc = app.Codec()
+	suite.cdc = app.AppCodec()
 
 	suite.now = time.Date(2020, 1, 2, 0, 0, 0, 0, time.UTC)
 	suite.privVal = tmtypes.NewMockPV()
@@ -46,6 +51,7 @@ func (suite *TendermintTestSuite) SetupTest() {
 	val := tmtypes.NewValidator(pubKey, 10)
 	suite.valSet = tmtypes.NewValidatorSet([]*tmtypes.Validator{val})
 	suite.header = ibctmtypes.CreateTestHeader(chainID, height, suite.now, suite.valSet, []tmtypes.PrivValidator{suite.privVal})
+	suite.ctx = app.BaseApp.NewContext(checkTx, abci.Header{Height: 1, Time: suite.now})
 }
 
 func TestTendermintTestSuite(t *testing.T) {