+++ x/ibc/09-localhost/types/client_state.go
@@ -14,7 +14,9 @@ import (
 	clientexported "github.com/cosmos/cosmos-sdk/x/ibc/02-client/exported"
 	clienttypes "github.com/cosmos/cosmos-sdk/x/ibc/02-client/types"
 	connectionexported "github.com/cosmos/cosmos-sdk/x/ibc/03-connection/exported"
+	connectiontypes "github.com/cosmos/cosmos-sdk/x/ibc/03-connection/types"
 	channelexported "github.com/cosmos/cosmos-sdk/x/ibc/04-channel/exported"
+	channeltypes "github.com/cosmos/cosmos-sdk/x/ibc/04-channel/types"
 	commitmentexported "github.com/cosmos/cosmos-sdk/x/ibc/23-commitment/exported"
 	commitmenttypes "github.com/cosmos/cosmos-sdk/x/ibc/23-commitment/types"
 	host "github.com/cosmos/cosmos-sdk/x/ibc/24-host"
@@ -25,16 +27,14 @@ var _ clientexported.ClientState = ClientState{}
 
 // ClientState requires (read-only) access to keys outside the client prefix.
 type ClientState struct {
-	store   sdk.KVStore
 	ID      string `json:"id" yaml:"id"`
 	ChainID string `json:"chain_id" yaml:"chain_id"`
 	Height  int64  `json:"height" yaml:"height"`
 }
 
 // NewClientState creates a new ClientState instance
-func NewClientState(store sdk.KVStore, chainID string, height int64) ClientState {
+func NewClientState(chainID string, height int64) ClientState {
 	return ClientState{
-		store:   store,
 		ID:      clientexported.Localhost.String(),
 		ChainID: chainID,
 		Height:  height,
@@ -68,19 +68,13 @@ func (cs ClientState) IsFrozen() bool {
 
 // Validate performs a basic validation of the client state fields.
 func (cs ClientState) Validate() error {
-	if err := host.DefaultClientIdentifierValidator(cs.ID); err != nil {
-		return err
-	}
 	if strings.TrimSpace(cs.ChainID) == "" {
 		return errors.New("chain id cannot be blank")
 	}
 	if cs.Height <= 0 {
 		return fmt.Errorf("height must be positive: %d", cs.Height)
 	}
-	if cs.store == nil {
-		return errors.New("KVStore cannot be nil")
-	}
-	return nil
+	return host.DefaultClientIdentifierValidator(cs.ID)
 }
 
 // VerifyClientConsensusState verifies a proof of the consensus
@@ -88,6 +82,7 @@ func (cs ClientState) Validate() error {
 // VerifyClientConsensusState verifies a proof of the consensus state of the
 // Tendermint client stored on the target machine.
 func (cs ClientState) VerifyClientConsensusState(
+	store sdk.KVStore,
 	cdc *codec.Codec,
 	_ commitmentexported.Root,
 	height uint64,
@@ -102,7 +97,7 @@ func (cs ClientState) VerifyClientConsensusState(
 		return err
 	}
 
-	data := cs.store.Get([]byte(path.String()))
+	data := store.Get([]byte(path.String()))
 	if len(data) == 0 {
 		return sdkerrors.Wrapf(clienttypes.ErrFailedClientConsensusStateVerification, "not found for path %s", path)
 	}
@@ -125,7 +120,8 @@ func (cs ClientState) VerifyClientConsensusState(
 // VerifyConnectionState verifies a proof of the connection state of the
 // specified connection end stored locally.
 func (cs ClientState) VerifyConnectionState(
-	cdc *codec.Codec,
+	store sdk.KVStore,
+	cdc codec.Marshaler,
 	_ uint64,
 	prefix commitmentexported.Prefix,
 	_ commitmentexported.Proof,
@@ -138,17 +134,18 @@ func (cs ClientState) VerifyConnectionState(
 		return err
 	}
 
-	bz := cs.store.Get([]byte(path.String()))
+	bz := store.Get([]byte(path.String()))
 	if bz == nil {
 		return sdkerrors.Wrapf(clienttypes.ErrFailedConnectionStateVerification, "not found for path %s", path)
 	}
 
-	var prevConnection connectionexported.ConnectionI
-	if err := cdc.UnmarshalBinaryBare(bz, &prevConnection); err != nil {
+	var prevConnection connectiontypes.ConnectionEnd
+	err = cdc.UnmarshalBinaryBare(bz, &prevConnection)
+	if err != nil {
 		return err
 	}
 
-	if connectionEnd != prevConnection {
+	if connectionEnd != &prevConnection {
 		return sdkerrors.Wrapf(
 			clienttypes.ErrFailedConnectionStateVerification,
 			"connection end ≠ previous stored connection: \n%v\n≠\n%v", connectionEnd, prevConnection,
@@ -161,7 +158,8 @@ func (cs ClientState) VerifyConnectionState(
 // VerifyChannelState verifies a proof of the channel state of the specified
 // channel end, under the specified port, stored on the local machine.
 func (cs ClientState) VerifyChannelState(
-	cdc *codec.Codec,
+	store sdk.KVStore,
+	cdc codec.Marshaler,
 	_ uint64,
 	prefix commitmentexported.Prefix,
 	_ commitmentexported.Proof,
@@ -175,16 +173,18 @@ func (cs ClientState) VerifyChannelState(
 		return err
 	}
 
-	bz := cs.store.Get([]byte(path.String()))
+	bz := store.Get([]byte(path.String()))
 	if bz == nil {
 		return sdkerrors.Wrapf(clienttypes.ErrFailedChannelStateVerification, "not found for path %s", path)
 	}
 
-	var prevChannel channelexported.ChannelI
-	if err := cdc.UnmarshalBinaryBare(bz, &prevChannel); err != nil {
+	var prevChannel channeltypes.Channel
+	err = cdc.UnmarshalBinaryBare(bz, &prevChannel)
+	if err != nil {
 		return err
 	}
-	if channel != prevChannel {
+
+	if channel != &prevChannel {
 		return sdkerrors.Wrapf(
 			clienttypes.ErrFailedChannelStateVerification,
 			"channel end ≠ previous stored channel: \n%v\n≠\n%v", channel, prevChannel,
@@ -197,6 +197,7 @@ func (cs ClientState) VerifyChannelState(
 // VerifyPacketCommitment verifies a proof of an outgoing packet commitment at
 // the specified port, specified channel, and specified sequence.
 func (cs ClientState) VerifyPacketCommitment(
+	store sdk.KVStore,
 	_ uint64,
 	prefix commitmentexported.Prefix,
 	_ commitmentexported.Proof,
@@ -211,7 +212,7 @@ func (cs ClientState) VerifyPacketCommitment(
 		return err
 	}
 
-	data := cs.store.Get([]byte(path.String()))
+	data := store.Get([]byte(path.String()))
 	if len(data) == 0 {
 		return sdkerrors.Wrapf(clienttypes.ErrFailedPacketCommitmentVerification, "not found for path %s", path)
 	}
@@ -229,6 +230,7 @@ func (cs ClientState) VerifyPacketCommitment(
 // VerifyPacketAcknowledgement verifies a proof of an incoming packet
 // acknowledgement at the specified port, specified channel, and specified sequence.
 func (cs ClientState) VerifyPacketAcknowledgement(
+	store sdk.KVStore,
 	_ uint64,
 	prefix commitmentexported.Prefix,
 	_ commitmentexported.Proof,
@@ -243,7 +245,7 @@ func (cs ClientState) VerifyPacketAcknowledgement(
 		return err
 	}
 
-	data := cs.store.Get([]byte(path.String()))
+	data := store.Get([]byte(path.String()))
 	if len(data) == 0 {
 		return sdkerrors.Wrapf(clienttypes.ErrFailedPacketAckVerification, "not found for path %s", path)
 	}
@@ -262,6 +264,7 @@ func (cs ClientState) VerifyPacketAcknowledgement(
 // incoming packet acknowledgement at the specified port, specified channel, and
 // specified sequence.
 func (cs ClientState) VerifyPacketAcknowledgementAbsence(
+	store sdk.KVStore,
 	_ uint64,
 	prefix commitmentexported.Prefix,
 	_ commitmentexported.Proof,
@@ -275,7 +278,7 @@ func (cs ClientState) VerifyPacketAcknowledgementAbsence(
 		return err
 	}
 
-	data := cs.store.Get([]byte(path.String()))
+	data := store.Get([]byte(path.String()))
 	if data != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedPacketAckAbsenceVerification, "expected no ack absence")
 	}
@@ -286,6 +289,7 @@ func (cs ClientState) VerifyPacketAcknowledgementAbsence(
 // VerifyNextSequenceRecv verifies a proof of the next sequence number to be
 // received of the specified channel at the specified port.
 func (cs ClientState) VerifyNextSequenceRecv(
+	store sdk.KVStore,
 	_ uint64,
 	prefix commitmentexported.Prefix,
 	_ commitmentexported.Proof,
@@ -299,7 +303,7 @@ func (cs ClientState) VerifyNextSequenceRecv(
 		return err
 	}
 
-	data := cs.store.Get([]byte(path.String()))
+	data := store.Get([]byte(path.String()))
 	if len(data) == 0 {
 		return sdkerrors.Wrapf(clienttypes.ErrFailedNextSeqRecvVerification, "not found for path %s", path)
 	}