+++ x/ibc/20-transfer/keeper/relay_test.go
@@ -5,8 +5,6 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/bank"
-	connectionexported "github.com/cosmos/cosmos-sdk/x/ibc/03-connection/exported"
-	channelexported "github.com/cosmos/cosmos-sdk/x/ibc/04-channel/exported"
 	channeltypes "github.com/cosmos/cosmos-sdk/x/ibc/04-channel/types"
 	"github.com/cosmos/cosmos-sdk/x/ibc/20-transfer/types"
 	ibctypes "github.com/cosmos/cosmos-sdk/x/ibc/types"
@@ -27,8 +25,8 @@ func (suite *KeeperTestSuite) TestSendTransfer() {
 			func() {
 				suite.chainA.App.BankKeeper.AddCoins(suite.chainA.GetContext(), testAddr1, testCoins)
 				suite.chainA.CreateClient(suite.chainB)
-				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, connectionexported.OPEN)
-				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, channelexported.OPEN, channelexported.ORDERED, testConnection)
+				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, ibctypes.OPEN)
+				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, ibctypes.OPEN, ibctypes.ORDERED, testConnection)
 				suite.chainA.App.IBCKeeper.ChannelKeeper.SetNextSequenceSend(suite.chainA.GetContext(), testPort1, testChannel1, 1)
 			}, true, true},
 		{"successful transfer from external chain", prefixCoins,
@@ -37,8 +35,8 @@ func (suite *KeeperTestSuite) TestSendTransfer() {
 				_, err := suite.chainA.App.BankKeeper.AddCoins(suite.chainA.GetContext(), testAddr1, prefixCoins)
 				suite.Require().NoError(err)
 				suite.chainA.CreateClient(suite.chainB)
-				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, connectionexported.OPEN)
-				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, channelexported.OPEN, channelexported.ORDERED, testConnection)
+				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, ibctypes.OPEN)
+				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, ibctypes.OPEN, ibctypes.ORDERED, testConnection)
 				suite.chainA.App.IBCKeeper.ChannelKeeper.SetNextSequenceSend(suite.chainA.GetContext(), testPort1, testChannel1, 1)
 			}, false, true},
 		{"source channel not found", testCoins,
@@ -46,32 +44,32 @@ func (suite *KeeperTestSuite) TestSendTransfer() {
 		{"next seq send not found", testCoins,
 			func() {
 				suite.chainA.CreateClient(suite.chainB)
-				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, connectionexported.OPEN)
-				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, channelexported.OPEN, channelexported.ORDERED, testConnection)
+				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, ibctypes.OPEN)
+				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, ibctypes.OPEN, ibctypes.ORDERED, testConnection)
 			}, true, false},
 		// createOutgoingPacket tests
 		// - source chain
 		{"send coins failed", testCoins,
 			func() {
 				suite.chainA.CreateClient(suite.chainB)
-				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, connectionexported.OPEN)
-				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, channelexported.OPEN, channelexported.ORDERED, testConnection)
+				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, ibctypes.OPEN)
+				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, ibctypes.OPEN, ibctypes.ORDERED, testConnection)
 				suite.chainA.App.IBCKeeper.ChannelKeeper.SetNextSequenceSend(suite.chainA.GetContext(), testPort1, testChannel1, 1)
 			}, true, false},
 		// - receiving chain
 		{"send from module account failed", testCoins,
 			func() {
 				suite.chainA.CreateClient(suite.chainB)
-				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, connectionexported.OPEN)
-				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, channelexported.OPEN, channelexported.ORDERED, testConnection)
+				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, ibctypes.OPEN)
+				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, ibctypes.OPEN, ibctypes.ORDERED, testConnection)
 				suite.chainA.App.IBCKeeper.ChannelKeeper.SetNextSequenceSend(suite.chainA.GetContext(), testPort1, testChannel1, 1)
 			}, false, false},
 		{"channel capability not found", testCoins,
 			func() {
 				suite.chainA.App.BankKeeper.AddCoins(suite.chainA.GetContext(), testAddr1, testCoins)
 				suite.chainA.CreateClient(suite.chainB)
-				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, connectionexported.OPEN)
-				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, channelexported.OPEN, channelexported.ORDERED, testConnection)
+				suite.chainA.createConnection(testConnection, testConnection, testClientIDB, testClientIDA, ibctypes.OPEN)
+				suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, ibctypes.OPEN, ibctypes.ORDERED, testConnection)
 				suite.chainA.App.IBCKeeper.ChannelKeeper.SetNextSequenceSend(suite.chainA.GetContext(), testPort1, testChannel1, 1)
 				// Release channel capability
 				cap, _ := suite.chainA.App.ScopedTransferKeeper.GetCapability(suite.chainA.GetContext(), capName)
@@ -130,7 +128,7 @@ func (suite *KeeperTestSuite) TestOnRecvPacket() {
 		// - receiving chain
 		{"incorrect dest prefix on coin denom",
 			func() {
-				data.Amount = prefixCoins2
+				data.Amount = prefixCoins
 			}, false},
 		{"success receive from external chain",
 			func() {
@@ -164,8 +162,7 @@ func (suite *KeeperTestSuite) TestOnRecvPacket() {
 // TestOnAcknowledgementPacket tests that successful acknowledgement is a no-op
 // and failure acknowledment leads to refund
 func (suite *KeeperTestSuite) TestOnAcknowledgementPacket() {
-	data := types.NewFungibleTokenPacketData(prefixCoins, testAddr1.String(), testAddr2.String())
-	testCoins2 := sdk.NewCoins(sdk.NewCoin("testportid/secondchannel/atom", sdk.NewInt(100)))
+	data := types.NewFungibleTokenPacketData(prefixCoins2, testAddr1.String(), testAddr2.String())
 
 	successAck := types.FungibleTokenPacketAcknowledgement{
 		Success: true,
@@ -186,13 +183,13 @@ func (suite *KeeperTestSuite) TestOnAcknowledgementPacket() {
 			func() {}, true, true},
 		{"successful refund from source chain", failedAck,
 			func() {
-				escrow := types.GetEscrowAddress(testPort2, testChannel2)
+				escrow := types.GetEscrowAddress(testPort1, testChannel1)
 				_, err := suite.chainA.App.BankKeeper.AddCoins(suite.chainA.GetContext(), escrow, sdk.NewCoins(sdk.NewCoin("atom", sdk.NewInt(100))))
 				suite.Require().NoError(err)
 			}, true, false},
 		{"successful refund from external chain", failedAck,
 			func() {
-				data.Amount = testCoins2
+				data.Amount = prefixCoins
 			}, false, false},
 	}
 
@@ -204,17 +201,17 @@ func (suite *KeeperTestSuite) TestOnAcknowledgementPacket() {
 		suite.Run(fmt.Sprintf("Case %s", tc.msg), func() {
 			suite.SetupTest() // reset
 
-			preCoin := suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), testAddr1, prefixCoins[0].Denom)
-
 			tc.malleate()
 
 			var denom string
 			if tc.source {
-				prefix := types.GetDenomPrefix(packet.GetSourcePort(), packet.GetSourceChannel())
-				denom = prefixCoins[0].Denom[len(prefix):]
+				prefix := types.GetDenomPrefix(packet.GetDestPort(), packet.GetDestChannel())
+				denom = prefixCoins2[0].Denom[len(prefix):]
 			} else {
 				denom = data.Amount[0].Denom
 			}
+
+			preCoin := suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), testAddr1, denom)
 
 			err := suite.chainA.App.TransferKeeper.OnAcknowledgementPacket(suite.chainA.GetContext(), packet, data, tc.ack)
 			suite.Require().NoError(err, "valid test case %d failed: %s", i, tc.msg)
@@ -225,7 +222,7 @@ func (suite *KeeperTestSuite) TestOnAcknowledgementPacket() {
 			if tc.success {
 				suite.Require().Equal(sdk.ZeroInt(), deltaAmount, "successful ack changed balance")
 			} else {
-				suite.Require().Equal(prefixCoins[0].Amount, deltaAmount, "failed ack did not trigger refund")
+				suite.Require().Equal(prefixCoins2[0].Amount, deltaAmount, "failed ack did not trigger refund")
 			}
 		})
 	}
@@ -233,8 +230,8 @@ func (suite *KeeperTestSuite) TestOnAcknowledgementPacket() {
 
 // TestOnTimeoutPacket test private refundPacket function since it is a simple wrapper over it
 func (suite *KeeperTestSuite) TestOnTimeoutPacket() {
-	data := types.NewFungibleTokenPacketData(prefixCoins, testAddr1.String(), testAddr2.String())
-	testCoins2 := sdk.NewCoins(sdk.NewCoin("testportid/secondchannel/atom", sdk.NewInt(100)))
+	data := types.NewFungibleTokenPacketData(prefixCoins2, testAddr1.String(), testAddr2.String())
+	testCoins2 := sdk.NewCoins(sdk.NewCoin("bank/firstchannel/atom", sdk.NewInt(100)))
 
 	testCases := []struct {
 		msg      string
@@ -244,7 +241,7 @@ func (suite *KeeperTestSuite) TestOnTimeoutPacket() {
 	}{
 		{"successful timeout from source chain",
 			func() {
-				escrow := types.GetEscrowAddress(testPort2, testChannel2)
+				escrow := types.GetEscrowAddress(testPort1, testChannel1)
 				_, err := suite.chainA.App.BankKeeper.AddCoins(suite.chainA.GetContext(), escrow, sdk.NewCoins(sdk.NewCoin("atom", sdk.NewInt(100))))
 				suite.Require().NoError(err)
 			}, true, true},
@@ -261,7 +258,7 @@ func (suite *KeeperTestSuite) TestOnTimeoutPacket() {
 			}, true, false},
 		{"mint failed",
 			func() {
-				data.Amount[0].Denom = prefixCoins[0].Denom
+				data.Amount[0].Denom = prefixCoins2[0].Denom
 				data.Amount[0].Amount = sdk.ZeroInt()
 			}, true, false},
 	}
@@ -274,17 +271,17 @@ func (suite *KeeperTestSuite) TestOnTimeoutPacket() {
 		suite.Run(fmt.Sprintf("Case %s", tc.msg), func() {
 			suite.SetupTest() // reset
 
-			preCoin := suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), testAddr1, prefixCoins[0].Denom)
-
 			tc.malleate()
 
 			var denom string
 			if tc.source {
-				prefix := types.GetDenomPrefix(packet.GetSourcePort(), packet.GetSourceChannel())
-				denom = prefixCoins[0].Denom[len(prefix):]
+				prefix := types.GetDenomPrefix(packet.GetDestPort(), packet.GetDestChannel())
+				denom = prefixCoins2[0].Denom[len(prefix):]
 			} else {
 				denom = data.Amount[0].Denom
 			}
+
+			preCoin := suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), testAddr1, denom)
 
 			err := suite.chainA.App.TransferKeeper.OnTimeoutPacket(suite.chainA.GetContext(), packet, data)
 
@@ -293,7 +290,7 @@ func (suite *KeeperTestSuite) TestOnTimeoutPacket() {
 
 			if tc.expPass {
 				suite.Require().NoError(err, "valid test case %d failed: %s", i, tc.msg)
-				suite.Require().Equal(prefixCoins[0].Amount.Int64(), deltaAmount.Int64(), "failed ack did not trigger refund")
+				suite.Require().Equal(prefixCoins2[0].Amount.Int64(), deltaAmount.Int64(), "successful timeout did not trigger refund")
 			} else {
 				suite.Require().Error(err, "invalid test case %d passed: %s", i, tc.msg)
 			}